Replaced Piano with new version
