--- src/components/PianoNew/Piano.scss
@@ -1,86 +1,0 @@
-.piano {
-  display: inline-flex;
-  padding: 0.125rem;
-  background-color: black;
-}
-
-.key {
-  --key-width: calc(1.73 * 4rem);
-
-  position: relative;
-  display: block;
-  width: var(--key-width);
-  height: 15rem;
-  background-color: white;
-  color: var(--text-dark-contrast-locked);
-  cursor: pointer;
-
-  &:hover {
-    &::after {
-      content: "";
-      position: absolute;
-      top: 0;
-      left: 0;
-      right: 0;
-      bottom: 0;
-      background-color: rgba(238, 212, 19, 0.2);
-    }
-  }
-
-  &:focus-within {
-    outline: solid rgb(238, 212, 19) 0.25rem;
-    outline-offset: -0.25rem;
-
-    &::after {
-      content: "";
-      position: absolute;
-      top: 0;
-      left: 0;
-      right: 0;
-      bottom: 0;
-      background-color: rgba(238, 212, 19, 0.3);
-    }
-  }
-
-  &__input {
-    position: absolute;
-    opacity: 0;
-    cursor: pointer;
-    height: 0;
-    width: 0;
-  }
-
-  &__label {
-    z-index: 1;
-    position: absolute;
-    bottom: 1rem;
-    left: 50%;
-    display: grid;
-    place-content: center;
-    width: 1.5rem;
-    height: 1.5rem;
-    border-radius: 50%;
-    font-size: 1rem;
-    transform: translateX(-50%);
-  }
-
-  &__input:checked + &__label {
-    background-color: red;
-    color: var(--text-light-contrast-locked);
-  }
-
-  &--sharp {
-    --key-width: calc(1 * 4rem);
-    
-    z-index: 1;
-    margin-left: calc(var(--key-width) * -1);
-    height: 10rem;
-    background-color: black;
-    color: var(--text-light-contrast-locked);
-    transform: translateX(50%);
-  }
-
-  &:not(:last-child, &--sharp) {
-    margin-right: 0.125rem;
-  }
-}
