--- src/components/PianoNew/index.ts
@@ -1,1 +1,0 @@
-export { default as Piano } from './Piano.svelte';
